Showing an error with custom rules.
In this commit I changed zodiac to use a custom rule. For the
purposes of example this rule just checks if the field isEmpty.

When you first blur the field or hit the validate button the field is
correctly marked as invalid. However, after this if you change the
field (and then empty it again) or even blur the field again then
the custom rule is never run for empty content (as can be verified
by the console.log statement).

I'm not sure what is happening here?