Introducing a property which only needs to be validated based on the status of another property.
When a user isNinja then they need to input the "secret ninja word"
which should be "NINJA!". Fields to toggle the isNinja status and
enter the secret word are now available on example 2.

This seems to work OK but breaks if you do the following:

 * Load the demo and navigate to example 2
 * Check "isNinja"
 * Enter something in the "secret ninja word" box then delete it
 * Uncheck "isNinja"
 * Click "Validate"

You will see the error message for "secret ninja word is required" is
shown even though the word shouldn't be required when isNinja is false.