Using `Ember.isEmpty` rather than a custom `isEmpty` method.
I was running into problems with the required rule not being applied
when a model was validated before you had visited all of the fields.

These turned out to be because of undefined values. `Ember.isEmpty`
catches `undefined` as well as `null` and `''` (and also an empty
array but I think that's OK?)